refactor importer package with trickle and balanced dag generation
